+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
  - docs: clarify that `bitround` and `zfp` codec configurations are draft
 
+### Fixed
+ - Do not serialise `configuration` in `Metadata` if is empty
+ - Do not serialise `endian` in `BytesCodecConfigurationV1` if it is none
+
 ## [0.11.1] - 2024-01-29
 
 ### Fixed